refactor(dsutil): remove dsutil package
pushing zip utilities into archive opens the door to moving all API helpers
up into the api package as private functions.

Replaced "UnmarshalDatasetYaml", with fill.Struct, which is a far better serializer
for reading from yaml. This function was only being used in FormFileDataset func.
