+++ api/root.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 
 	util "github.com/qri-io/apiutil"
-	"github.com/qri-io/qri/base/dsfs/dsutil"
 	"github.com/qri-io/qri/config"
 	"github.com/qri-io/qri/fsi"
 	"github.com/qri-io/qri/lib"
@@ -77,7 +76,7 @@ func (mh *RootHandler) Handler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if err := dsutil.InlineScriptsToBytes(res.Dataset); err != nil {
+	if err := inlineScriptsToBytes(res.Dataset); err != nil {
 		util.WriteErrResponse(w, http.StatusInternalServerError, err)
 		return
 	}
