+++ api/testdata/complete/body.csv
@@ -1,0 +1,6 @@
+city,pop,avg_age,in_usa
+toronto,40000000,55.5,false
+new york,8500000,44.4,true
+chicago,300000,44.4,true
+chatham,35000,65.25,true
+raleigh,250000,50.65,true
