+++ api/testdata/complete/rendered.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>dataset with all submodels example</title>
+  </head>
+  <body>
+    <h1>dataset with all submodels example</h1>
+  </body>
+</html>
