+++ api/testdata/complete/template.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>{{ .Meta.Title }}</title>
+</head>
+<body>
+  <h1>{{ .Meta.Title }}</h1>
+</body>
+</html>
