+++ api/testdata/complete/transform.star
@@ -1,0 +1,1 @@
+commit([1,2,3,4,5])
